A program that adds fibonacci nums that are divisible by 2.

+++ 0x02-functions_nested_loops/103-fibonacci.c
@@ -8,7 +8,7 @@
 int main(void)
 {
 	int i;
-	int sum = 2;
+	long int sum = 2;
 	long int fib[50];
 
 	fib[0] = 1;
